Allow `x_band`, `x_lines`, `x_minrange` for any plot type (specifically scatter) (#2851)
* Allow x_band, x_lines, x_minrange functionality for any plot type (specifically Scatter)

* Fix warnings

* Temporarily disregard generic type values in get_x_range

* Fix passing cls to parse_ methods in validation

* Fix

* Fix
